+++ tests/unit/Double/StubContainerConfig.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 namespace WoohooLabs\Zen\Tests\Unit\Double;
 
 use WoohooLabs\Zen\Config\AbstractContainerConfig;
+use WoohooLabs\Zen\Config\Hint\DefinitionHintInterface;
+use WoohooLabs\Zen\Config\Hint\WildcardHintInterface;
 
 class StubContainerConfig extends AbstractContainerConfig
 {
@@ -12,9 +14,21 @@ class StubContainerConfig extends AbstractContainerConfig
      */
     private $entryPoints;
 
-    public function __construct(array $entryPoints)
+    /**
+     * @var DefinitionHintInterface[]
+     */
+    private $definitionHints;
+
+    /**
+     * @var WildcardHintInterface[]
+     */
+    private $wildcardHints;
+
+    public function __construct(array $entryPoints = [], array $definitionHints = [], array $wildcardHints = [])
     {
         $this->entryPoints = $entryPoints;
+        $this->definitionHints = $definitionHints;
+        $this->wildcardHints = $wildcardHints;
     }
 
     protected function getEntryPoints(): array
@@ -24,11 +38,11 @@ class StubContainerConfig extends AbstractContainerConfig
 
     protected function getDefinitionHints(): array
     {
-        return [];
+        return $this->definitionHints;
     }
 
     protected function getWildcardHints(): array
     {
-        return [];
+        return $this->wildcardHints;
     }
 }